label-print-service/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pillow>=10.2.0
brother-ql>=0.9.4
pyusb>=1.2.1
